Fix nb tests
